+++ packages/moengage_flutter/moengage_flutter_web/lib/constants.dart
@@ -9,12 +9,17 @@ const String methodLogout = 'logout';
 
 // SDK Method Names
 const String methodTrackEventSDK = 'track_event';
+const String methodSetIntegrationTypeSDK = 'setIntegrationType';
 const String methodSetUserAttributeSDK = 'add_user_attribute';
 const String methodSetUniqueIdSDK = 'add_unique_user_id';
 const String methodSetAliasSDK = 'update_unique_user_id';
+const String methodIdentifyUserSDK = 'identifyUser';
+const String methodGetUserIdentitiesSDK = 'getUserIdentities';
 const String methodLogoutSDK = 'destroy_session';
 const String methodEnableSDK = 'enableSdk';
 const String methodDisableSDK = 'disableSdk';
+const String methodEnableDataTracking = 'enableDataTracking';
+const String methodDisableDataTracking = 'disableDataTracking';
 
 // Keys Constants
 String keyEventName = 'eventName';